clean data, list rendering defined in App
