+++ src/components/SearchList/SearchList.jsx
@@ -1,16 +1,8 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchSearch } from '../../redux/actions';
+import React from 'react';
+import { useSelector } from 'react-redux';
 
 function SearchList() {
-  const dispatch = useDispatch();
-  const query = useSelector((store) => store.query);
-  const { data } = useSelector((store) => store.data);
-
-  useEffect(() => {
-    const queryText = query.text.split(' ').join('%20');
-    dispatch(fetchSearch(queryText, query.entitie));
-  }, [query]);
+  const data = useSelector((store) => store.data.data);
 
   return (
     <>
@@ -26,11 +18,6 @@ function SearchList() {
           </ol>
         )
       }
-      {/* {data.artists && (
-        <ol>
-          {data.artists.map((art) => <li key={art.id}>{art.name}</li>)}
-        </ol>
-      )} */}
     </>
   );
 }
