+++ src/redux/actions/index.js
@@ -3,12 +3,14 @@ import fetchSearchAPI from '../../services/musicBrainz';
 export const ADD_ARTIST = 'ADD_ARTIST';
 export const ADD_QUERY = 'ADD_QUERY';
 export const ADD_DATA = 'ADD_DATA';
+export const CLEAN_DATA = 'CLEAN_DATA';
 
 export const addArtist = (artistID) => ({ type: ADD_ARTIST, id: artistID });
 
 export const addQuery = (text, entitie) => ({ type: ADD_QUERY, text, entitie });
 
 export const addData = (data) => ({ type: ADD_DATA, data });
+export const cleanData = () => ({ type: CLEAN_DATA });
 
 export const fetchSearch = (query, entitie) => async (dispatch) => {
   try {
