+++ src/redux/reducers/data.jsx
@@ -1,4 +1,4 @@
-import { ADD_DATA } from '../actions';
+import { ADD_DATA, CLEAN_DATA } from '../actions';
 
 const INITIAL = {
   data: {},
@@ -7,7 +7,9 @@ const INITIAL = {
 export default function data(state = INITIAL, action) {
   switch (action.type) {
     case ADD_DATA:
-      return { data: action.data };
+      return { ...state, data: action.data };
+    case CLEAN_DATA:
+      return { ...state, data: {} };
     default:
       return state;
   }
